pyproject.toml
[tool.poetry]
name = "drf_shoestore"
version = "0.1.0"
description = ""
authors = ["Ben McKenzie <bmckenzie0522@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.1"
djangorestframework = "^3.12.0"
django_mptt = "^0.11.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
